Only clean up the model api if it exists.

+++ jujugui/static/gui/src/app/init.js
@@ -1611,7 +1611,7 @@ class GUIApp {
       clearTimeout(this._dbChangedTimer);
     }
     // Destroy YUI classes.
-    this.modelAPI.destroy();
+    this.modelAPI && this.modelAPI.destroy();
     this.controllerAPI.destroy();
     this.db.destroy();
     this.endpointsController.destroy();
